stats

+++ project4/src/header_files/global_variables.h
@@ -3,7 +3,6 @@
 
 extern int totalEntitiesInHierarchy;
 extern int totalEntitiesCopied;
-extern int totalBytesCopied;
-
+extern long long int totalBytesCopied;
 
 #endif
